+++ backend/src/main/java/hanglog/global/config/ResolverConfig.java
@@ -1,6 +1,6 @@
-package hanglog.config;
+package hanglog.global.config;
 
-import hanglog.global.jwt.AuthArgumentResolver;
+import hanglog.auth.AuthArgumentResolver;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
@@ -9,7 +9,7 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
 @RequiredArgsConstructor
-public class WebMvcConfig implements WebMvcConfigurer {
+public class ResolverConfig implements WebMvcConfigurer {
 
     private final AuthArgumentResolver authArgumentResolver;
